+++ Packages/io.chainsafe.web3-unity/Runtime/Scripts/EVM/Remote.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a29b9c314e7272d4daaed8afb17e28e7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
