+++ Packages/io.chainsafe.web3-unity/Runtime/Scripts/EVM/Remote/CSServer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ea43c720228be347a5ecebba0c5a83e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
